% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeTreeData.R
\docType{class}
\name{bridgeTreeData-class}
\alias{bridgeTreeData-class}
\alias{bridgeTreeData}
\title{bridgeTree Data Class}
\description{
A class for managing metadata associated with bridgeTree queries and data manipulation.
}
\section{Slots}{

\describe{
\item{\code{userQueryType}}{A vector of type character specifying whether the user made their original taxonomic query based on a vector of taxon names or a phylogeny.}

\item{\code{userSpecTaxonomicSources}}{A vector of type character that presents a list of taxonomic sources for cleaning taxonomy of queries. This can be user-specified or default.}

\item{\code{cleanedTaxonomy}}{A dataframe with containing input taxon names, the closeset match according to \code{\link{gnr_resolve}}, and a list of taxonomic datasources that contain the matching name, generated by \code{\link{studyTaxonList}}.}

\item{\code{occSources}}{A vector of class "character" containing a list of occurrence data sources, generated when passing a \code{\link{bridgeTreeData}} object through \code{\link{taxaQuery}}.}

\item{\code{occNLimit}}{The limit of occurrence records to return for each species when executing \code{\link{occ}} through \code{\link{taxaQuery}}.}

\item{\code{occurrenceSearchDate}}{The date on which the occurrence search query was conducted.}

\item{\code{occResults}}{The results of an \code{\link{taxQuery}} search, stored as a named list, each of the items named after a searched taxon and containing a dataframe with occurrence information.}
}}

