% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occQuery.R
\name{occQuery}
\alias{occQuery}
\title{Query from Taxon List}
\usage{
occQuery(x = NULL, datasources = "gbif", GBIFLogin = NULL,
  options = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{bridgeTreeData}} (the results of a \code{\link{studyTaxonList}} search).}

\item{datasources}{A vector of occurrence datasources to search. This is currently limited to GBIF, but may expand in the future.}

\item{GBIFLogin}{An object of class \code{\link{GBIFLogin}} to log in to GBIF to begin the download.}

\item{options}{A vector of options to pass to \code{\link{occ_download}}.}
}
\value{
The object of class \code{\link{bridgeTreeData}} supplied by the user as an argument, with occurrence data search results, as well as metadata on the occurrence sources queried.
}
\description{
Takes rectified list of specimens from \code{\link{studyTaxonList}} and returns point data from \code{\link{rgbif}} with metadata.
}
\examples{
## PLACEHOLDER
studyTaxonList(x = phylogeny, datasources = c('NCBI', 'EOL'));

## PLACEHOLDER
studyTaxonList(x = c("Buteo buteo", "Buteo buteo hartedi", "Buteo japonicus"), datasources = c('NCBI', 'EOL'));

}
