% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyTaxonList.R
\name{studyTaxonList}
\alias{studyTaxonList}
\title{Study Taxon Vector}
\usage{
studyTaxonList(x)
}
\arguments{
\item{x}{}
}
\value{
nameVector
}
\description{
Takes input phylogenies or vectors of taxon names, checks against taxonomic database, returns vector for use in spocc queries, as well as warnings if there are invalid names.
}
\examples{
## Inputting a phylogeny
studyTaxonList(x = phylogeny);

## Inputting a vector of taxon names
studyTaxonList(x = c("Buteo buteo", "Buteo buteo hartedi", "Buteo japonicus"));

}
