% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaQuery.R
\name{taxaQuery}
\alias{taxaQuery}
\title{Query from Taxon List}
\usage{
taxaQuery(x = NULL, datasources = "gbif", limit = 500, options = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{bridgeTreeData}} (the results of a \code{\link{studyTaxonList}} search).}

\item{datasources}{A vector of occurrence datasources to search, as implemented in \code{\link{occ}}.}

\item{limit}{The maximum number of datapoints that should be returned from an \code{\link{occ}} search.}

\item{options}{A vector of options to pass to \code{\link{occ}}.}
}
\value{
pointsList A list containing results from each species' \code{\link{occ}} search, including metadata.
}
\description{
Takes rectified list of specimens from \code{\link{studyTaxonList}} and returns point data from \code{\link{occ}} with metadata.
}
\examples{
## PLACEHOLDER
studyTaxonList(x = phylogeny, datasources = c('NCBI', 'EOL'));

## PLACEHOLDER
studyTaxonList(x = c("Buteo buteo", "Buteo buteo hartedi", "Buteo japonicus"), datasources = c('NCBI', 'EOL'));

}
